+++ main.py
@@ -3,6 +3,7 @@ from juegos.JuegoAdivinaAnimal import iniciar_juego
 from juegos.ChOrD3 import juego_musica
 from juegos.tablas_de_multiplicar import LlamarJuegoMultiplicar
 from juegos.ultra_septimum import jugar
+from juegos.preguntas8_python import jugar_8preguntas
 menu = 0
 
 while menu != 8:
@@ -18,6 +19,7 @@ while menu != 8:
     print("        |3| →   Tablas de multiplicar     |")
     print("        |4| →            Ch0rD3           |")
     print("        |5| →        Ultra Septimum       |")
+    print("        |6| →     8 Preguntas Python      |")
     print("        |8| →            SALIR            |")
     print()
     menu = int(input("              Opcion a elegir →"))
@@ -31,6 +33,8 @@ while menu != 8:
         juego_musica()
     elif menu == 5:
         jugar()
+    elif menu == 6:
+        jugar_8preguntas()
     elif menu == 8:
         print("            PROGRAMA FINALIZADO")
